REF: Rename module pca -> principal_components
