+++ statsmodels/tools/principal_components.py
